+++ addons/simple_cards/card/card_layout/card_layout.gd
@@ -3,15 +3,18 @@
 ##[color=green]Use the "Card Layouts" tab at the bottom for a quick setup![/color][br]
 ##Alone is pretty much useless. Use other nodes to build your best layout for the cards that need to be made.[br]
 ##The Subviewport required by this node will also give the card size. (Subviewport.size = card.size)
+@tool
 @icon("uid://dfeet1bp3au3l")
 class_name CardLayout extends SubViewportContainer
 
-
-signal layout_ready ##Emitted at the end of the setup function.
+##Emitted from _layout_ready() after ready-time initialization.
+signal layout_ready()
 ##Emitted when layout is initialized
 signal layout_initialized(card: Card, resource: CardResource)
 ##Emitted after display is updated
 signal display_updated()
+##Emitted when [member card_size] changes. Connected by [Card] to reactively update its size.
+signal card_size_changed(new_size: Vector2i)
 ##Emitted when flip in animation starts
 signal flip_in_started()
 ##Emitted when flip in animation completes
@@ -28,6 +31,16 @@ signal focus_in_completed()
 signal focus_out_started()
 ##Emitted when focus out animation completes
 signal focus_out_completed()
+##Emitted at the end of setup()
+signal setup_completed(card: Card, resource: CardResource)
+
+##Size of the card in pixels. Updates the [SubViewport] and notifies the parent [Card] when changed.
+##[br]Defaults to [code]Vector2i.ZERO[/code] which auto-initializes from the SubViewport's size on [method _ready].
+@export_custom(PROPERTY_HINT_NONE, "suffix:px") var card_size: Vector2i = Vector2i.ZERO:
+	set(value):
+		if card_size == value: return
+		card_size = value
+		_apply_card_size()
 
 
 ##Reference to the [CardResource] used to initialize the layout. Use it to further customize the layout.
@@ -79,13 +92,33 @@ func _focus_out() -> void:
 	focus_out_started.emit()
 	if focus_out_animation:
 		await focus_out_animation.play_animation(self)
+
 	focus_out_completed.emit()
+## Triggered at the end of [code]_ready()[/code]. [color=red]Override[/color] it to implement custom behaviour.
+func _layout_ready() -> void:
+	layout_ready.emit()
+
+func _ready() -> void:
+	if card_size == Vector2i.ZERO:
+		var sub_vp = get_node_or_null("SubViewport")
+		if sub_vp and sub_vp is SubViewport and sub_vp.size != Vector2i.ZERO:
+			card_size = sub_vp.size
+	_layout_ready()
+
+## Syncs the child [SubViewport] size to [member card_size] and emits [signal card_size_changed].
+func _apply_card_size() -> void:
+	if !is_inside_tree(): return
+	var sub_vp = get_node_or_null("SubViewport")
+	if sub_vp and sub_vp is SubViewport:
+		sub_vp.size = card_size
+	card_size_changed.emit(card_size)
 
 
-##Used when a [Card] initializes a layout, set the reference of the card its resource and emits [member CardLayout.layout_ready] at the end.
+##Used when a [Card] initializes a layout, sets the card/resource references, updates the display,
+##and emits [member CardLayout.setup_completed] at the end.
 func setup(card: Card, resource: CardResource) -> void:
 	card_instance = card
 	card_resource = resource
 	_update_display()
 	layout_initialized.emit(card, resource)
-	layout_ready.emit()
+	setup_completed.emit(card, resource)